server/requirements.txt
asgiref==3.2.10
cbauth==1.0.7
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.8
django-allauth==0.42.0
django-cors-headers==3.4.0
django-filter==2.3.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
idna==2.10
install==1.3.3
Markdown==3.2.2
oauthlib==3.1.0
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
sqlparse==0.3.1
urllib3==1.25.9
